Commit 17 - Minor changed in domain classes and completed another test in Distance test class

+++ src/main/java/com/miticorp/topography/basic/model/DistanceTypeMetricCentimeter.java
@@ -1,7 +1,7 @@
 package com.miticorp.topography.basic.model;
 
 public class DistanceTypeMetricCentimeter extends DistanceTypeMetric {
-	private Double conversionToMeter = (double) (1 / (100));
+	private Double conversionToMeter = 1D / 100;
 
 	@Override
 	public Double getConversionToMeter() {
