+++ src/main/java/com/miticorp/topography/basic/model/DistanceTypeMetricMillimeter.java
@@ -1,7 +1,7 @@
 package com.miticorp.topography.basic.model;
 
 public class DistanceTypeMetricMillimeter extends DistanceTypeMetric {
-	private Double conversionToMeter = (double) (1 / (1000));
+	private Double conversionToMeter = 1D / 1000;
 
 	@Override
 	public Double getConversionToMeter() {
